buld the ans vocab

+++ classifier/evaluate.py
@@ -36,7 +36,7 @@ def evaluate(data_iter, model, TEXT, emb_dim, LABELS, VAR_VALUES, ANS, snis, pre
         # True Acc
         parser = Parser()
         var_values = np.array(VAR_VALUES.vocab.itos)[np.array(batch.var_values.data)]
-        ans = np.array(ANS.vocab.itos)[np.array(batch.ans.data)]
+        ans = np.array(ANS.vocab.itos)[np.array(batch.ans.data)].astype(float)
 
         pred_answers = []
         tgt_answers = []
